Expression: Move `StrCompareMixin` to separate `mixins` sub-paackge
