+++ loki/expression/__init__.py
@@ -9,9 +9,10 @@ Expression layer of the two-level Loki IR based on `Pymbolic
 <https://github.com/inducer/pymbolic>`_.
 """
 
-from loki.expression.symbols import *  # noqa
-from loki.expression.operations import *  # noqa
-from loki.expression.mappers import *  # noqa
-from loki.expression.symbolic import *  # noqa
-from loki.expression.parser import *  # noqa
 from loki.expression.evaluation import *  # noqa
+from loki.expression.mappers import *  # noqa
+from loki.expression.mixins import *  # noqa
+from loki.expression.operations import *  # noqa
+from loki.expression.parser import *  # noqa
+from loki.expression.symbolic import *  # noqa
+from loki.expression.symbols import *  # noqa
